back-end/requirements.txt
asgiref==3.2.10
certifi==2020.12.5
decorator==4.4.2
Django==3.1.2
django-3-jet==1.0.7
django-ckeditor==6.0.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-js-asset==1.2.2
django-versatileimagefield==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.1.2
intervaltree==3.1.0
libmagic==1.0
networkx==2.5.1
pep8==1.7.1
Pillow==8.2.0
psycopg2==2.8.6;platform_system=="Windows"
PyJWT==2.0.1
python-dateutil==2.8.1
python-magic==0.4.22
python-magic-bin==0.4.14;platform_system=="Windows"
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
tag==0.5
text-unidecode==1.3
uritemplate==3.0.1
django-cors-headers
psycopg2-binary
mypy
pycodestyle
django-colorfield
parameterized
djangorestframework-camel-case